# Scratch JSON changes to investigate various techniques to improve performance