--- a/tests/Benchmarks/System.IO.Pipelines/E2E.cs
+++ b/tests/Benchmarks/System.IO.Pipelines/E2E.cs
@@ -2,7 +2,9 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Buffers;
 using System.Buffers.Text;
+using System.Buffers.Writer;
 using System.IO.Pipelines.Samples;
 using System.Text;
 using System.Text.Formatting;
@@ -55,9 +57,9 @@ namespace System.IO.Pipelines.Benchmarks
                 formatter.Format("\r\nDate: {0:R}", DateTime.UtcNow);
                 formatter.Append("Server: System.IO.Pipelines");
                 formatter.Append("\r\n\r\n");
-
+                
                 // write body
-                var writer = new JsonWriter(formatter, true);
+                var writer = new JsonWriter(new BufferWriter<IBufferWriter<byte>>(formatter), true, true);
                 writer.WriteObjectStart();
                 writer.WriteAttribute("message", "Hello, World!");
                 writer.WriteObjectEnd();