Related to the discussions in https://github.com/dotnet/corefxlab/pull/2358 / https://github.com/dotnet/corefxlab/pull/2366 and builds on top of https://github.com/dotnet/corefxlab/pull/2358.

I don't intend to merge this in as is, but wanted to share the source and performance results of trying different approaches when using BufferWriter APIs. Here are the different performance tests (_note what the user ends up writing for the "Compact" versions - this changes what JsonWriter APIs we provide_).

**HelloWorld**
```C#
json.WriteObjectStart();
json.WriteAttribute("message", "Hello, World!");
json.WriteObjectEnd();
```

**HelloWorldCompact**
```C#
json.WriteObject("message", "Hello, World!");
```

**ArrayValues**
```C#
json.WriteArrayStart("ExtraArray");
for (var i = 0; i < data.Length; i++)
{
    json.WriteValue(data[i]);
}
json.WriteArrayEnd();
```
**ArrayValuesCompact**
```C#
json.WriteArray("ExtraArray", data);
```

**(A)** Master (non-generic JsonWriter without BufferWriter)
**(B)** Generic JsonWriter (with BufferWriter) without using BufferWriter specific APIs
**(C)** Generic JsonWriter (with BufferWriter) without using BufferWriter specific APIs but using TryWrite pattern rather than computing required space ahead of time (see https://github.com/dotnet/corefxlab/issues/2294)
**(D)** Generic JsonWriter (with BufferWriter) using BufferWriter convenient Write APIs
**(E)** Generic JsonWriter  (with BufferWriter) using custom IFormattable structs and BufferWriter TWritable APIs

**Conclusions:**

- For HelloWorld, A and C are best and E is worst. Compact APIs always help, so we should explore that further.
- For ArrayValues, D is best and E is worst. B scales similar to D. For ArrayValuesCompact, C and E are best and D is worst. A is better for small array sizes (less than 10), but C scales better.
- For ArrayValues (pretty), B is best and E is worst. For ArrayValuesCompact (pretty), C and E are best and D is worst. A is better for small array sizes (less than 10), but C scales better.

The delta between A and B is caused by the overhead of making JsonWriter generic and storing BufferWriter as a field (roughly 25% for HelloWorld). C is generally on par with A and scales better. We should apply the optimizations for C regardless. D is on part with B but doesn't benefit for the Compact APIs. E benefits from Compact APIs, but otherwise is the slowest. We don't get much benefit from using custom IFormattable structs. There is overhead of using the Write APIs which we may be able to reduce. **The best performance comes from not using BufferWriter APIs at all, and keeping JsonWriter non-generic (i.e. A with improvements from C).**


cc @benaadams, @KrzysztofCwalina, @JeremyKuhne 

![image](https://user-images.githubusercontent.com/6527137/41697040-e33dc30a-74cb-11e8-9b2e-4c28951676d3.png)

``` ini

BenchmarkDotNet=v0.10.14.534-nightly, OS=Windows 10.0.14393.2312 (1607/AnniversaryUpdate/Redstone1)
Intel Core i7-6700 CPU 3.40GHz (Skylake), 1 CPU, 8 logical and 4 physical cores
Frequency=3328123 Hz, Resolution=300.4697 ns, Timer=TSC
.NET Core SDK=2.2.100-preview1-009007
  [Host]     : .NET Core 2.2.0-preview1-26529-01 (CoreCLR 4.6.26525.04, CoreFX 4.6.26525.01), 64bit RyuJIT
  DefaultJob : .NET Core 2.2.0-preview1-26529-01 (CoreCLR 4.6.26525.04, CoreFX 4.6.26525.01), 64bit RyuJIT


```

**(A) Master (non-generic JsonWriter without BufferWriter):**

|                                Method | Formatted |      Mean |     Error |    StdDev |
|-------------------------------------- |---------- |----------:|----------:|----------:|
|        **WriterSystemTextJsonHelloWorld** |     **False** | **115.45 ns** | **2.2834 ns** | **2.3448 ns** |
| WriterSystemTextJsonHelloWorldCompact |     False |  77.39 ns | 0.2353 ns | 0.2201 ns |
|        **WriterSystemTextJsonHelloWorld** |      **True** | **118.05 ns** | **0.7711 ns** | **0.7213 ns** |
| WriterSystemTextJsonHelloWorldCompact |      True |  86.72 ns | 0.5523 ns | 0.5166 ns |

|                               Method | Formatted | ExtraArraySize |          Mean |        Error |       StdDev |
|------------------------------------- |---------- |--------------- |--------------:|-------------:|-------------:|
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **1** |     **102.50 ns** |    **17.118 ns** |    **11.322 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              1 |      67.19 ns |     8.419 ns |     5.010 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **2** |     **127.82 ns** |    **15.782 ns** |    **10.439 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              2 |      84.30 ns |    14.352 ns |     8.541 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |             **10** |     **266.43 ns** |     **7.796 ns** |     **4.639 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |             10 |     167.98 ns |    14.503 ns |     9.593 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **100** |   **2,267.02 ns** |   **150.648 ns** |    **99.644 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            100 |   1,137.97 ns |    66.782 ns |    44.172 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **500** |  **11,310.70 ns** |   **467.593 ns** |   **309.284 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            500 |   6,053.05 ns |   188.227 ns |   124.501 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |           **5000** | **110,314.97 ns** | **1,745.271 ns** | **1,154.389 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |           5000 |  91,103.49 ns | 2,286.111 ns | 1,195.682 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **1** |      **99.01 ns** |    **20.134 ns** |    **11.981 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              1 |      76.05 ns |     7.288 ns |     4.337 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **2** |     **117.54 ns** |     **9.441 ns** |     **6.245 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              2 |     101.19 ns |    24.770 ns |    16.384 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |             **10** |     **273.25 ns** |    **18.449 ns** |    **10.979 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |             10 |     203.48 ns |     9.840 ns |     5.856 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **100** |   **2,454.20 ns** |   **106.441 ns** |    **70.404 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            100 |   1,672.68 ns |   132.670 ns |    87.753 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **500** |  **11,772.41 ns** |   **102.946 ns** |    **68.092 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            500 |   8,651.76 ns |   101.975 ns |    67.450 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |           **5000** | **116,995.55 ns** |   **777.910 ns** |   **514.539 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |           5000 | 106,373.58 ns |   208.670 ns |   109.138 ns |

**(B) Generic JsonWriter without using BufferWriter specific APIs:**

|                                Method | Formatted |     Mean |     Error |    StdDev |
|-------------------------------------- |---------- |---------:|----------:|----------:|
|        **WriterSystemTextJsonHelloWorld** |     **False** | **146.9 ns** | **3.2677 ns** | **4.2489 ns** |
| WriterSystemTextJsonHelloWorldCompact |     False | 116.3 ns | 3.1863 ns | 3.1294 ns |
|        **WriterSystemTextJsonHelloWorld** |      **True** | **149.0 ns** | **0.4082 ns** | **0.3409 ns** |
| WriterSystemTextJsonHelloWorldCompact |      True | 119.8 ns | 0.6131 ns | 0.5735 ns |

|                               Method | Formatted | ExtraArraySize |         Mean |        Error |       StdDev |
|------------------------------------- |---------- |--------------- |-------------:|-------------:|-------------:|
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **1** |     **121.3 ns** |    **18.383 ns** |    **12.159 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              1 |     101.6 ns |     8.966 ns |     5.930 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **2** |     **131.7 ns** |    **12.906 ns** |     **8.536 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              2 |     119.6 ns |    14.181 ns |     9.380 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |             **10** |     **256.4 ns** |    **30.461 ns** |    **20.148 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |             10 |     189.7 ns |     6.042 ns |     3.160 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **100** |   **1,814.3 ns** |    **81.655 ns** |    **54.010 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            100 |   1,184.0 ns |    80.343 ns |    47.811 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **500** |   **9,054.3 ns** |   **187.240 ns** |   **123.848 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            500 |   6,908.5 ns |   629.726 ns |   374.740 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |           **5000** |  **94,365.9 ns** |   **924.504 ns** |   **611.503 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |           5000 |  90,566.4 ns | 3,031.071 ns | 1,803.742 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **1** |     **122.5 ns** |    **14.251 ns** |     **9.426 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              1 |     131.4 ns |    24.233 ns |    16.028 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **2** |     **144.4 ns** |    **20.874 ns** |    **13.807 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              2 |     129.4 ns |    19.577 ns |    12.949 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |             **10** |     **279.9 ns** |    **31.820 ns** |    **18.936 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |             10 |     248.6 ns |    16.389 ns |     9.753 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **100** |   **2,080.6 ns** |    **91.845 ns** |    **60.749 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            100 |   1,615.7 ns |   140.248 ns |    83.459 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **500** |  **10,369.4 ns** |   **125.995 ns** |    **83.338 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            500 |   8,714.7 ns |   324.169 ns |   192.908 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |           **5000** | **106,614.5 ns** |   **581.417 ns** |   **304.093 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |           5000 | 106,574.5 ns | 3,217.593 ns | 2,128.239 ns |

**(C) Generic JsonWriter without using BufferWriter specific APIs but using TryWrite pattern rather than computing required space ahead of time (see https://github.com/dotnet/corefxlab/issues/2294):**

|                                Method | Formatted |      Mean |     Error |    StdDev |
|-------------------------------------- |---------- |----------:|----------:|----------:|
|        **WriterSystemTextJsonHelloWorld** |     **False** | **116.52 ns** | **1.3205 ns** | **1.1027 ns** |
| WriterSystemTextJsonHelloWorldCompact |     False |  95.01 ns | 0.6199 ns | 0.5799 ns |
|        **WriterSystemTextJsonHelloWorld** |      **True** | **110.65 ns** | **0.6026 ns** | **0.5637 ns** |
| WriterSystemTextJsonHelloWorldCompact |      True | 100.88 ns | 0.8466 ns | 0.7919 ns |

|                               Method | Formatted | ExtraArraySize |          Mean |        Error |       StdDev |
|------------------------------------- |---------- |--------------- |--------------:|-------------:|-------------:|
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **1** |     **110.44 ns** |     **9.881 ns** |     **4.387 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              1 |      87.53 ns |     9.994 ns |     6.610 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **2** |     **122.46 ns** |     **9.438 ns** |     **5.617 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              2 |      99.70 ns |     6.130 ns |     4.055 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |             **10** |     **296.07 ns** |    **43.701 ns** |    **26.006 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |             10 |     151.29 ns |    17.517 ns |    11.586 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **100** |   **2,204.17 ns** |   **120.399 ns** |    **71.648 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            100 |     910.63 ns |   107.492 ns |    71.100 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **500** |  **11,508.51 ns** |   **555.230 ns** |   **330.409 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            500 |   4,431.07 ns |    59.546 ns |    31.144 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |           **5000** | **117,507.85 ns** | **2,790.272 ns** | **1,845.592 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |           5000 |  58,159.08 ns | 1,048.444 ns |   623.912 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **1** |     **113.25 ns** |    **17.803 ns** |    **11.776 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              1 |     101.87 ns |    19.073 ns |    11.350 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **2** |     **138.65 ns** |    **27.022 ns** |    **17.873 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              2 |     109.30 ns |    12.591 ns |     7.493 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |             **10** |     **321.36 ns** |    **33.605 ns** |    **22.228 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |             10 |     211.43 ns |    29.065 ns |    19.225 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **100** |   **2,488.97 ns** |   **109.375 ns** |    **72.345 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            100 |   1,301.27 ns |    72.412 ns |    47.896 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **500** |  **12,510.09 ns** |    **79.677 ns** |    **47.415 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            500 |   6,617.26 ns |   214.526 ns |   141.896 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |           **5000** | **131,577.99 ns** | **4,116.600 ns** | **2,449.724 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |           5000 |  73,325.98 ns |   631.079 ns |   417.419 ns |

**(D) Generic JsonWriter using BufferWriter convenient Write APIs:**

|                                Method | Formatted |     Mean |     Error |    StdDev |
|-------------------------------------- |---------- |---------:|----------:|----------:|
|        **WriterSystemTextJsonHelloWorld** |     **False** | **122.2 ns** | **2.8639 ns** | **2.8127 ns** |
| WriterSystemTextJsonHelloWorldCompact |     False | 118.0 ns | 0.3587 ns | 0.3355 ns |
|        **WriterSystemTextJsonHelloWorld** |      **True** | **155.1 ns** | **0.5664 ns** | **0.4730 ns** |
| WriterSystemTextJsonHelloWorldCompact |      True | 147.3 ns | 0.5673 ns | 0.5306 ns |

|                               Method | Formatted | ExtraArraySize |         Mean |        Error |       StdDev |
|------------------------------------- |---------- |--------------- |-------------:|-------------:|-------------:|
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **1** |     **102.3 ns** |    **15.639 ns** |     **9.306 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              1 |     109.9 ns |     6.950 ns |     4.597 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |              **2** |     **124.3 ns** |    **27.706 ns** |    **16.487 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |              2 |     124.5 ns |     5.497 ns |     2.875 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |             **10** |     **221.1 ns** |    **21.987 ns** |    **13.084 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |             10 |     334.0 ns |    35.822 ns |    23.694 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **100** |   **1,749.2 ns** |   **104.860 ns** |    **62.401 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            100 |   3,016.6 ns |    94.713 ns |    62.647 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |            **500** |   **8,492.8 ns** |   **182.142 ns** |   **120.475 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |            500 |  14,844.2 ns |   260.914 ns |   172.578 ns |
|        **WriterSystemTextJsonArrayOnly** |     **False** |           **5000** |  **86,186.4 ns** | **1,452.547 ns** |   **864.388 ns** |
| WriterSystemTextJsonArrayOnlyCompact |     False |           5000 | 146,071.4 ns |   561.439 ns |   249.284 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **1** |     **138.8 ns** |    **20.325 ns** |    **13.444 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              1 |     163.5 ns |    18.333 ns |     9.589 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |              **2** |     **169.6 ns** |    **24.084 ns** |    **15.930 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |              2 |     207.5 ns |    25.171 ns |    14.979 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |             **10** |     **354.2 ns** |    **39.317 ns** |    **23.397 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |             10 |     570.7 ns |    54.460 ns |    32.408 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **100** |   **2,802.4 ns** |    **96.051 ns** |    **63.532 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            100 |   4,657.7 ns |   113.831 ns |    75.292 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |            **500** |  **13,132.5 ns** |   **231.755 ns** |   **137.914 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |            500 |  24,614.2 ns | 2,540.127 ns | 1,680.137 ns |
|        **WriterSystemTextJsonArrayOnly** |      **True** |           **5000** | **134,261.4 ns** | **2,810.383 ns** | **1,858.894 ns** |
| WriterSystemTextJsonArrayOnlyCompact |      True |           5000 | 225,358.9 ns | 1,405.313 ns |   836.279 ns |

**(E) Generic JsonWriter using custom IFormattable structs and BufferWriter TWritable APIs:**

|                                Method | Formatted |     Mean |     Error |    StdDev |
|-------------------------------------- |---------- |---------:|----------:|----------:|
|        **WriterSystemTextJsonHelloWorld** |     **False** | **163.9 ns** | **2.8098 ns** | **2.3463 ns** |
| WriterSystemTextJsonHelloWorldCompact |     False | 113.5 ns | 1.3147 ns | 1.2297 ns |
…