@benaadams 2018-06-21
> best performance comes from not using BufferWriter APIs at all...

To clarify, is this using `BufferWriter<T>`; but not the `BufferWriter<T>` apis, like `Write(int)` etc?