@ahsonkhan 2018-06-21
> > best performance comes from not using BufferWriter APIs at all...

> To clarify, is this using `BufferWriter<T>`; but not the `BufferWriter<T>` apis, like `Write(int)` etc?

The comment is about not using `BufferWriter<T>` at all (and keeping JsonWriter non-generic, with only an `IBufferWriter<byte>` field). I have alternatives where we use BufferWriter but don't use APIs like `Write(int)`, etc.