@benaadams 2018-06-21
> > > best performance comes from not using BufferWriter APIs at all...

> > To clarify, is this using `BufferWriter<T>`; but not the `BufferWriter<T>` apis, like Write(int) etc?

> The comment is about not using `BufferWriter<T>` at all 

It would have more effect for a struct writer and something with a more expensive `GetSpan`/`Advance` though, for example the `BufferWriterFormatter<PipeWriter>`?