@AndyAyersMS 2018-06-21
I meant to look at what I think you are calling the A-B perf diff but got sidetracked by dotnet/coreclr#18522 and some other stuff. Think that is now mostly out of the way so with any luck I can look deeper tomorrow morning.

I find it unfortunate that to gain deeper insight I often have to move the code out of BDN so I can it under run more typical profilers like perfview -- @adamsitnik says some kind of ETL based profiling is on the way for BDN, but in the meantime it takes me longer than I'd like to analyze stuff.