@benaadams 2018-06-21
The benchmark tests are numbered 0 - N in the command line column in the order they show up in the table:

![image](https://user-images.githubusercontent.com/1142958/41727324-f99a30d4-756b-11e8-8ffc-a1d8d64d01f2.png)

Can also reduce any `[Params]` attributes to a single entry; and comment out any additional `[Benchmark]` attributes for items you aren't interested in, to make it quicker.

Though I only started doing this over the weekend :)