@benaadams 2018-06-21
If I need to get something more detailed like Jit dumps; then I'll change the Program Main to create the Benchmark directly; rather than running a Benchmark e.g.

```csharp
public static void Main(string[] args)
{
    //BenchmarkSwitcher
    //    .FromAssemblyAndTypes(typeof(Program).Assembly, GetGenericBenchmarks())         
    //    .Run(args/*, CreateClrVsCoreConfig() uncomment it to run Clr vs .NET Core comparison*/);

    // Create benchmark
    var p = new JsonWriterPerf();
    // Set [Params] fields
    p.IsUTF8Encoded = true;
    p.Formatted = false;
    // Run [GlobalSetup]
    p.Setup();

    // Execute [Benchmark] Method
    p.WriterSystemTextJsonHelloWorld();
}
```